Taking a Break from Problem 1
Hi Jack!  I'm not sure exactly what is required to convert each player's classType into a class-- right now they're strings (which I know is not the most efficient way to think about solving this but in retrospect I'm not sure if I have as much cs background as I thought I did when signing up for this class)!  Thank you for being patient as always :)  I spent way too long on the first question as per usual.  I feel like if I truly limited myself to 30 minutes each I wouldn't be learning on my own, but I couldn't get the first one to work so far and I'm going to move onto the second problem now! As for the regular assignment, I'm debating whether I want to use the given Roguelike or if I want to do something super simple with my virtual DAW game, but I'll have to make that decision soon!